+++ quickstart/docker/image/ziti-cli-functions.sh
@@ -74,7 +74,7 @@ function stopController {
   if [[ -n ${ZITI_EXPRESS_CONTROLLER_PID:-} ]]; then
     kill "$ZITI_EXPRESS_CONTROLLER_PID"
     # shellcheck disable=SC2181
-    if [[ $? == 0 ]]; then 
+    if [[ $? == 0 ]]; then
       echo "Controller stopped."
       return 0
     fi
@@ -89,7 +89,7 @@ function startRouter {
   "${ZITI_BIN_DIR}/ziti-router" run "${ZITI_HOME_OS_SPECIFIC}/${ZITI_EDGE_ROUTER_RAWNAME}.yaml" > "${log_file}" 2>&1 &
   ZITI_EXPRESS_EDGE_ROUTER_PID=$!
   echo -e "Express Edge Router started as process id: $ZITI_EXPRESS_EDGE_ROUTER_PID. log located at: $(BLUE "${log_file}")"
-  
+
 }
 
 function stopRouter {
@@ -586,6 +586,10 @@ function ziti_expressConfiguration {
   "${ZITI_BIN_DIR-}/ziti-router" enroll "${ZITI_HOME_OS_SPECIFIC}/${ZITI_EDGE_ROUTER_RAWNAME}.yaml" --jwt "${ZITI_HOME_OS_SPECIFIC}/${ZITI_EDGE_ROUTER_RAWNAME}.jwt" &> "${ZITI_HOME_OS_SPECIFIC}/${ZITI_EDGE_ROUTER_RAWNAME}.enrollment.log"
   echo ""
 
+  echo -e "******** Setting Up Default Service ********"
+
+  createAPIService
+
   stopController
   echo "Edge Router enrolled. Controller stopped."
 
@@ -1076,7 +1080,7 @@ function ziti_createEnvFile {
   echo "ZITI_NETWORK set to: ${ZITI_NETWORK}"
 
   if [[ "${ZITI_USER-}" == "" ]]; then export ZITI_USER="admin"; fi
-  if [[ "${ZITI_PWD-}" == "" ]]; then 
+  if [[ "${ZITI_PWD-}" == "" ]]; then
     ZITI_PWD="$(LC_ALL=C tr -dc _A-Z-a-z-0-9 < /dev/urandom | head -c32)"
     echo -en "Do you want to keep the generated admin password '$ZITI_PWD'? (Y/n) "
     # shellcheck disable=SC2162
@@ -1550,6 +1554,25 @@ function getFileOverwritePermission() {
 
     return 0
   fi
+}
+
+function createAPIService {
+  echo "----------  Creating a REST API service ziti.rest.api:80...."
+
+  zitiLogin
+  # Create the bind and dial configs
+  "${ZITI_BIN_DIR-}/ziti" edge create config ziti.rest.dial intercept.v1 '{"protocols":["tcp"],"addresses":["ziti.rest.api"], "portRanges":[{"low":443, "high":443}]}' > /dev/null
+  "${ZITI_BIN_DIR-}/ziti" edge create config ziti.rest.bind host.v1 '{"protocol":"tcp", "address":"'${ZITI_EDGE_CONTROLLER_HOSTNAME}'","port":'"${ZITI_EDGE_CONTROLLER_PORT}"'}' > /dev/null
+
+  # Create the service, applying the configs
+  "${ZITI_BIN_DIR-}/ziti" edge create service ziti.rest.service --configs "ziti.rest.bind,ziti.rest.dial" > /dev/null
+
+  # Create the service policies allowing dial and bind access
+  "${ZITI_BIN_DIR-}/ziti" edge create service-policy ziti.rest.service.bind Bind --service-roles "@ziti.rest.service" --identity-roles "#ziti.rest.binders" > /dev/null
+  "${ZITI_BIN_DIR-}/ziti" edge create service-policy ziti.rest.service.dial Dial --service-roles "@ziti.rest.service" --identity-roles "#ziti.rest.dialers" > /dev/null
+
+  # Allow the express edge router to bind
+  "${ZITI_BIN_DIR-}/ziti" edge update identity "${ZITI_EDGE_ROUTER_RAWNAME}" -a ziti.rest.binders > /dev/null
 }
 
 set +uo pipefail